+++ cpp/src/custring_view.cu
@@ -1070,8 +1070,8 @@ __device__ custring_view* custring_view::replace(unsigned int pos, unsigned int 
     char* sptr = (char*)(this->data());
     custring_view* str = create_from(mem);
     char* optr = str->data();
-    if((pos >= end) || // range overlaps itself
-        (pos > nchars)) // outside of string's range
+    if((pos > end) ||  // range overlaps itself
+       (pos > nchars)) // outside of string's range
     {
         memcpy(optr, sptr, sz);
         str->init_fields(sz);
@@ -1110,8 +1110,8 @@ __device__ custring_view* custring_view::replace(unsigned int pos, unsigned int 
     char* sptr = this->data();
     custring_view* str = create_from(mem);
     char* optr = str->data();
-    if((pos >= end) || // range overlaps itself
-        (pos > nchars)) // outside of string's range
+    if((pos > end) ||  // range overlaps itself
+       (pos > nchars)) // outside of string's range
     {
         memcpy(optr, sptr, sz);
         str->init_fields(sz);
@@ -1152,8 +1152,8 @@ __device__ unsigned int custring_view::replace_size(unsigned int pos, unsigned i
     unsigned int end = pos + length;
     if(end > chars) // end is non-inclusize
         end = chars;
-    if((pos >= end) || // range overlaps itself
-        (pos > chars))  // outside of string's range
+    if((pos > end) ||  // range overlaps itself
+       (pos > chars))  // outside of string's range
         return alloc_size();
     //
     unsigned int left = offset_for_char_pos(pos);
@@ -1173,8 +1173,8 @@ __device__ unsigned int custring_view::replace_size(unsigned int pos, unsigned i
     unsigned int end = pos + length;
     if(end > chars) // end is non-inclusize
         end = chars;
-    if((pos >= end) || // range overlaps itself
-        (pos > chars))  // outside of string's range
+    if((pos > end) ||  // range overlaps itself
+       (pos > chars))  // outside of string's range
         return alloc_size();
     //
     unsigned int left = offset_for_char_pos(pos);
@@ -1620,25 +1620,7 @@ __device__ unsigned int custring_view::rstrip_size(const char* tgts) const
 // these expect only numbers (ascii charset)
 __device__ int custring_view::stoi() const
 {
-    char* ptr = (char*)data();
-    if(!ptr)
-        return 0; // probably should be an assert
-    int value = 0, sign = 1;
-    unsigned int sz = size();
-    if(*ptr == '-' || *ptr == '+')
-    {
-        sign = (*ptr == '-' ? -1 : 1);
-        ++ptr;
-        --sz;
-    }
-    for(unsigned int idx = 0; idx < sz; ++idx)
-    {
-        char ch = *ptr++;
-        if(ch < '0' || ch > '9')
-            break;
-        value = (value * 10) + (int)(ch - '0');
-    }
-    return value * sign;
+    return (int)stol();
 }
 
 __device__ long custring_view::stol() const
@@ -1660,7 +1642,7 @@ __device__ long custring_view::stol() const
         char ch = *ptr++;
         if(ch < '0' || ch > '9')
             break;
-        value = (value * 10) + (long)(ch - '0');
+        value = (value * 10L) + (long)(ch - '0');
     }
     return value * (long)sign;
 }
@@ -1685,19 +1667,23 @@ __device__ unsigned long custring_view::stoul() const
 
 __device__ float custring_view::stof() const
 {
+    return (float)stod();
+}
+
+__device__ double custring_view::stod() const
+{
     char* ptr = (char*)data();
     if(!ptr)
-        return 0.0f; // probably should be an assert
+        return 0.0; // probably should be an assert
     unsigned int sz = size();
     if( sz==0 )
-        return 0.0f;
+        return 0.0;
     char* end = ptr + sz;
-    float sign = 1;
+    double sign = 1.0;
     if(*ptr == '-' || *ptr == '+')
     {
         sign = (*ptr == '-' ? -1 : 1);
         ++ptr;
-        --sz;
     }
     unsigned int digits = 0;
     int exp_off = 0;
@@ -1743,39 +1729,53 @@ __device__ float custring_view::stof() const
     exp10 *= exp_sign;
     exp10 += exp_off;
     double value = (double)digits * pow(10.0,(double)exp10);
-    return (float)(value * sign);
+    return (value * sign);
 }
 
-__device__ double custring_view::stod() const
+__device__ custring_view* custring_view::ltos( long value, void* mem )
 {
-    char* ptr = (char*)data();
-    if(!ptr)
-        return 0.0; // probably should be an assert
+    if( value==0 )
+        return create_from(mem,"0",1);
+    char* str = (char*)mem;
+    char* ptr = str;
+    bool sign = value < 0;
+    if( sign )
+        value = -value;
+    while( value > 0 )
+    {
+        char ch = '0' + (value % 10);
+        *ptr++ = ch;
+        value = value/10;
+    }
+    if( sign )
+        *ptr++ = '-';
+    // number is backwards, so let's reverse it
+    int len = (int)(ptr-str);
+    for( int j=0; j<(len/2); ++j )
+    {
+        char ch1 = str[j];
+        char ch2 = str[len-j-1];
+        str[j] = ch2;
+        str[len-j-1] = ch1;
+    }
+    return create_from(str,str,len);
+}
 
-    double value = 0, factor = 1;
-    unsigned int sz = size();
-    if(*ptr == '-' || *ptr == '+')
+__device__ unsigned int custring_view::ltos_size( long value )
+{
+    if( value==0 )
+        return alloc_size(1,1);
+    bool sign = value < 0;
+    if( sign )
+        value = -value;
+    int digits = 0; // count the digits
+    while( value > 0 )
     {
-        factor = (*ptr == '-' ? -1 : 1);
-        ++ptr;
-        --sz;
+        ++digits;
+        value = value/10;
     }
-    bool decimal = false;
-    for(unsigned int idx = 0; idx < sz; ++idx)
-    {
-        char ch = *ptr++;
-        if(ch == '.')
-        {
-            decimal = true;
-            continue;
-        }
-        if(ch < '0' || ch > '9')
-            break;
-        if(decimal)
-            factor /= 10.0;
-        value = value * 10.0 + (double)(ch - '0'); // see float above
-    }
-    return value * factor;
+    int bytes = digits + (int)sign;
+    return alloc_size(bytes,bytes);
 }
 
 //